Feature/signupdetail

See merge request s09-fintech-finance-sub2/S09P22B207!14

+++ Front-End/pro/app/src/main/java/com/example/project/BartercreatePage.kt
@@ -29,7 +29,7 @@ fun BarterCreatePage(navController: NavHostController) {
     var currentPage by remember { mutableStateOf(1) }
     val itemsPerPage = 10
 
-    var selectedItemIndices by remember { mutableStateOf(listOf<Int>()) }
+    var selectedItemIndices by remember { mutableStateOf(listOf<Long>()) }
 
     Box(modifier = Modifier.fillMaxSize()) {
         Column(
